+++ core/Console/ClearCache.php
@@ -12,12 +12,13 @@ declare(strict_types=1);
 namespace Brocooly\Console;
 
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class ClearCache extends Command
 {
+	use SymfonyStyleTrait;
+
 	/**
 	 * The name of the command
 	 *
@@ -39,14 +40,11 @@ class ClearCache extends Command
 	 * @inheritDoc
 	 */
 	protected function execute( InputInterface $input, OutputInterface $output ) {
-
-		$io = new SymfonyStyle( $input, $output );
-
 		$loader = new \Timber\Loader();
 		$loader->clear_cache_twig();
 		$loader->clear_cache_timber();
 
-		$io->success( 'Cache was successfully flushed' );
+		$this->io->success( 'Cache was successfully flushed' );
 		return Command::SUCCESS;
 	}